crypto/ec/ec2_lopezdahab.c: Prototypal Lopez-Dahab ADD and DBL
This commit adds ``crypto/ec/ec2_lopezdahab.c`` to the tree, which
contains:

1. the code to initialize and finalize Lopez-Dahab context, i.e. the
   data structure used to compute ADD and DBL in Lopez-Dahab projective
   coordinates;

2. the (autogenerated) code to compute ADD and DBL in Lopez-Dahab
   coordinates;

3. the code to convert from/to Lopez-Dahab coordinates;

4. wrappers for EC_POINT_add() and EC_POINT_dbl() that receive params
   in affine coordinates, compute the result in Lopez-Dahab coordinates
   and convert it back to affine coordinates.

The code is autogenerated using the explicit formulae in "three operand
code" that can be downloaded at <http://www.hyperelliptic.org/EFD/g12o/>.
A subsequent commit will add the script used to autogenerate the code.

Wrappers are there to hijack ordinary GF(2m) code and run the tests, and
I expect them to go away when we will provide a MUL implementation in Lopez-
Dahab coordinates.