crypto/ec: minimal set of changes for testing Lopez-Dahab
This commit contains the minimal set of changes required to allow for
testing of the new Lopez-Dahab code.  The idea is to hijack the standard
double and add operations to wrappers that perform operations in Lopez-
Dahab coordinates but accept affine input and return affine output.

I needed to add an ABORT to ``crypto/ec/ectest.c`` because after that
point the code enters into an infinite loop.  I will investigate the
issue in the coming days and remove the ABORT as soon as I understand
the infinite loop cause.