apps/speed.c: Do not assume RSA support is available
This commit modifies apps/speed.c so that it does not assume
that RSA support is available:

1. it defines FOO_speed for FOO in dsa, ecdh, ecdsa, which is
   compiled in only when algorithm FOO is enabled;

2. it surrounds with #ifndef ... #endif a piece of code that
   must be compiled in only when RSA is enabled.

Additionally, when doing 2, it removes a duplicate piece of
code.

Works for me in the following cases:

a. ecdsa, ecdh enabled;

b. ecdsa enabled;

c. ecdsa, rsa enabled;

d. ecdsa, dsa enabled.

At the moment I'm not able to disable ecdsa, because it seems
another patch is needed.