ec2_lopezdahab.c: Fix DBL and ADD corner cases
We provide implementation for the corner cases when the input
is in affine coordinates only.  We assume that, when the input
is already in Lopez-Dahab, we are invoked by the MUL() and we
assume that, in this case, the caller has passed us a good point
and we will never get to the point at infinite.

Discussed with Emanuele.