"Everything together"
A little bit messy commit that fixes nearly all the problems and makes wNAF to properly work.
it is possible to appreciate a performance boost in the meaningful operations in ectest. Unfortunately, as ectest performs a lot of operations, the performance boost is only about 20%. A more precise benchmark must be developed