Removed internal optimizations
Now, even when the lopezdahab functions are called internally, all the checks are performed. This is necessary because, in this way, the add and dbl functions will always properly work. As I plan to re-use existing pieces of code to implement wNAF, this is necessary as it is not possible to guarantee that the wNAF code will perform these checks by itself